requirements.txt
pillow
pyocr
deep-translator
pyperclip
